# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Helpdesk API - REST",
        "description": "API REST del Sistema Helpdesk. Migración de GraphQL a REST con autenticación JWT.",
        "contact": {
            "name": "API Support",
            "email": "support@helpdesk.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/api/companies/followed": {
            "get": {
                "tags": [
                    "Company Followers"
                ],
                "summary": "List companies followed by user",
                "description": "Returns all companies that the authenticated user is following, ordered by most recent follow first. Includes company details and user-specific metrics like ticket count.",
                "operationId": "list_followed_companies",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of companies followed by the authenticated user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "company": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "format": "uuid"
                                                            },
                                                            "company_code": {
                                                                "type": "string"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "logo_url": {
                                                                "type": "string",
                                                                "format": "uri",
                                                                "nullable": true
                                                            },
                                                            "status": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "followed_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "my_tickets_count": {
                                                        "type": "integer"
                                                    },
                                                    "last_ticket_created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "has_unread_announcements": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/companies/{company}/is-following": {
            "get": {
                "tags": [
                    "Company Followers"
                ],
                "summary": "Check if user follows a company",
                "description": "Verifies if the authenticated user is currently following the specified company.",
                "operationId": "check_if_following",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "Company UUID or identifier",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Follow status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "isFollowing": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Company not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/companies/{company}/follow": {
            "post": {
                "tags": [
                    "Company Followers"
                ],
                "summary": "Follow a company",
                "description": "Allows the authenticated user to start following a company. If already following, returns current follow status. Rate limited to 20 requests per hour.",
                "operationId": "follow_company",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "Company UUID or identifier",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Already following the company",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "company": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "company_code": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "logo_url": {
                                                    "type": "string",
                                                    "format": "uri",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "followed_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "isFollowing": {
                                            "type": "boolean"
                                        },
                                        "followersCount": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Successfully started following the company",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "company": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "company_code": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "logo_url": {
                                                    "type": "string",
                                                    "format": "uri",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "followed_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "isFollowing": {
                                            "type": "boolean"
                                        },
                                        "followersCount": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Company not found"
                    },
                    "409": {
                        "description": "Already following this company"
                    },
                    "429": {
                        "description": "Too many follow requests (rate limit exceeded)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/companies/{company}/unfollow": {
            "delete": {
                "tags": [
                    "Company Followers"
                ],
                "summary": "Unfollow a company",
                "description": "Allows the authenticated user to stop following a company. Returns an error if the user is not currently following the company.",
                "operationId": "unfollow_company",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "description": "Company UUID or identifier",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully unfollowed the company",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "isFollowing": {
                                            "type": "boolean"
                                        },
                                        "followersCount": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Company not found"
                    },
                    "409": {
                        "description": "User is not following this company"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Bearer Token. Include in Authorization header: Bearer <token>",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}